+++ .env.sample
@@ -4,6 +4,6 @@ TAPYRUS_RPC_PORT=12381
 TAPYRUS_RPC_USER=rpcuser
 TAPYRUS_RPC_PASSWORD=rpcpassword
 TAPYRUS_AUTHORITY_KEY=cUJN5RVzYWFoeY8rUztd47jzXCu1p57Ay8V7pqCzsBD3PEXN7Dd4
-ISSUER_URL=http://localhost:3001
-ACQUIRER_URL=http://localhost:3002
+ISSUER_URL=http://localhost:3001/api
+ACQUIRER_URL=http://localhost:3002/api
 DID_SERVICE_URL=http://localhost:4000
